[Feat]#3-RegisterInputView 생성
