+++ hwProject/Model/UserManager.swift
@@ -28,7 +28,7 @@ class UserManager {
         }
     }
     
-    // 사용자를 추가합니다
+    // 사용자를 추가
     func addUser(email: String, password: String) {
         let user = User()
         user.email = email
@@ -39,17 +39,15 @@ class UserManager {
         }
     }
     
-    // 이메일로 사용자를 가져옵니다
     func getUser(byEmail email: String) -> User? {
         return realm.objects(User.self).filter("email == %@", email).first
     }
     
-    // 이메일이 등록되어 있는지 확인합니다
+    // 등론된 이메일인지 확인
     func isEmailRegistered(_ email: String) -> Bool {
         return realm.objects(User.self).filter("email == %@", email).count > 0
     }
     
-    // 이메일과 비밀번호로 사용자를 검증합니다
     func verifyUser(email: String, password: String) -> Bool {
         if let user = realm.objects(User.self).filter("email == %@ AND password == %@", email, password).first {
             loggedInUser = user
@@ -58,19 +56,16 @@ class UserManager {
         return false
     }
     
-    // 모든 사용자를 가져옵니다
     func getAllUsers() -> Results<User> {
         return realm.objects(User.self)
     }
     
-    // 사용자를 삭제합니다
     func deleteUser(user: User) {
         try! realm.write {
             realm.delete(user)
         }
     }
     
-    // 모든 사용자 정보를 출력합니다
     func printAllUsers() {
         let users = getAllUsers()
         users.forEach { user in
@@ -78,17 +73,14 @@ class UserManager {
         }
     }
     
-    // 로그인된 사용자를 가져옵니다
     func getLoggedInUser() -> User? {
         return loggedInUser
     }
     
-    // 로그인된 사용자를 설정합니다
     func setLoggedInUser(email: String) {
         loggedInUser = getUser(byEmail: email)
     }
     
-    // 로그아웃합니다
     func logoutUser() {
         loggedInUser = nil
     }
